// generated/src/app/api-models/profsvc-claim-detail-meas.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProfsvcClaimDetailMeas {
  seqClaimId: number;
  claimLineNumber: number;
  subLineCode: string;
  measLineNumber: number;
  orgDosage: string;
  testResultValue: string;
  measType: string;
  measValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}